requirements.txt
altair==4.2.0
pandas==1.3.5
Pillow==9.2.0
plotly==5.10.0
streamlit==1.12.0
vega_datasets==0.9.0
~treamlit==1.12.0
